move 'features' directory to 'routes' to work with remix project structure
we're not planning on migrating to Remix (give the news that remix is essentually being merged into react router v7) however, given this news, incrementally adopting more of the react reouter v6 features may help us take advantage of some of the useful features of remix when react router v7 comes out
