--- client/app/features/ErrorPage/index.ts
@@ -1,1 +1,0 @@
-export { ErrorPage as Component } from 'features/ErrorPage/ErrorPage';
