+++ client/app/routes/Dashboard/Dashboard.spec.tsx
@@ -1,5 +1,5 @@
 import '@testing-library/jest-dom';
-import { Dashboard } from 'features/Dashboard/Dashboard';
+import { Dashboard } from 'routes/Dashboard/Dashboard';
 import { setupServer } from 'msw/node';
 import React, { createElement } from 'react';
 import { cleanup, renderWithProviders, screen } from 'test-utils';
