+++ client/app/routes/ErrorPage/index.ts
@@ -1,0 +1,1 @@
+export { ErrorPage as Component } from 'routes/ErrorPage/ErrorPage';
