+++ client/app/routes/ManifestDetails/ManifestDetails.tsx
