+++ client/app/routes/ManifestDetails/index.ts
