+++ client/app/routes/RegisterHero/RegisterHero.tsx
