+++ client/app/routes/SiteDetails/index.ts
@@ -1,4 +1,4 @@
-import { SiteDetails } from 'features/SiteDetails/SiteDetails';
+import { SiteDetails } from 'routes/SiteDetails/SiteDetails';
 
 export { SiteDetails as Component };
 export default SiteDetails;
