+++ client/app/routes/SiteList/index.ts
@@ -1,4 +1,4 @@
-import { SiteList } from 'features/SiteList/SiteList';
+import { SiteList } from 'routes/SiteList/SiteList';
 
 export { SiteList as Component };
 export default SiteList;
